backend_form/requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.8.3
requests==2.31.0
python-dotenv==1.0.1
